+++ 1.21.10/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.10.2-1.21.10] - 2025-11-15
+
+### Changed
+
+- All enchantment compatibility changes are now controlled via separate built-in data packs
+- Changes to bow, crossbow and mace enchantments remain enabled by default
+- Changes to damage and protection enchantments remain disabled by default
+
 ## [v21.10.1-1.21.10] - 2025-10-23
 
 ### Changed